server/requirements.txt
requests
flask
flask_cors
python-dotenv
# Link preview
linkpreview
lxml
beautifulsoup4